+++ assets/src/apps/authoring/components/RightMenu/RightMenu.tsx
@@ -3,9 +3,10 @@ import { Tab, Tabs } from 'react-bootstrap';
 import { useSelector } from 'react-redux';
 import { selectCurrentActivity } from '../../../delivery/store/features/activities/slice';
 import { selectState as selectPageState } from '../../store/page/slice';
-import PropertyEditor, { Schema } from '../PropertyEditor/PropertyEditor';
-import lessonSchema from '../PropertyEditor/schemas/lesson';
-import screenSchema from '../PropertyEditor/schemas/screen';
+import PropertyEditor from '../PropertyEditor/PropertyEditor';
+import lessonSchema, {lessonUiSchema} from '../PropertyEditor/schemas/lesson';
+import screenSchema, {getScreenData, screenUiSchema} from '../PropertyEditor/schemas/screen';
+import { JSONSchema7 } from 'json-schema';
 
 const RightMenu: React.FC<any> = (props) => {
   const [selectedTab, setSelectedTab] = useState<string>('lesson');
@@ -13,14 +14,23 @@ const RightMenu: React.FC<any> = (props) => {
   const currentLesson = useSelector(selectPageState);
 
   // TODO: dynamically load schema from Part Component configuration
-  const componentSchema: Schema = { type: 'object' };
+  const componentSchema: JSONSchema7 = { type: 'object' };
   const currentComponent = null;
-
+  const screenData = getScreenData(currentActivity?.model.custom);
   const handleSelectTab = (key: string) => {
     // TODO: any other saving or whatever
     setSelectedTab(key);
   };
 
+  const screenPropertyChangeHandler = (properties:any) => {
+    //console.log(properties);
+  };
+  const lessonPropertyChangeHandler = (properties:any) => {
+    console.log(properties);
+  };
+  const componentPropertyChangeHandler = (properties:any) => {
+    //console.log(properties);
+  };
   return (
     <Tabs
       className="aa-panel-section-title-bar aa-panel-tabs"
@@ -28,13 +38,30 @@ const RightMenu: React.FC<any> = (props) => {
       onSelect={handleSelectTab}
     >
       <Tab eventKey="lesson" title="Lesson">
-        <PropertyEditor schema={lessonSchema} value={currentLesson} />
+        <PropertyEditor
+          schema={lessonSchema}
+          uiSchema={lessonUiSchema}
+          value={currentLesson}
+          onChangeHandler={lessonPropertyChangeHandler}
+        />
       </Tab>
       <Tab eventKey="screen" title="Screen">
-        <PropertyEditor schema={screenSchema} value={currentActivity?.model.custom} />
+        <PropertyEditor
+          schema={screenSchema}
+          uiSchema={screenUiSchema}
+          value={screenData}
+          onChangeHandler={screenPropertyChangeHandler}
+        />
       </Tab>
       <Tab eventKey="component" title="Component" disabled={!currentComponent}>
-        {currentComponent && <PropertyEditor schema={componentSchema} value={currentComponent} />}
+        {currentComponent && (
+          <PropertyEditor
+            schema={componentSchema}
+            uiSchema={{}}
+            value={currentComponent}
+            onChangeHandler={componentPropertyChangeHandler}
+          />
+        )}
       </Tab>
     </Tabs>
   );