+++ assets/src/apps/delivery/store/features/adaptivity/actions/triggerCheck.ts
@@ -1,21 +1,21 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import { RootState } from 'apps/delivery/store/rootReducer';
 import { PartResponse } from 'components/activities/types';
-import { evalActivityAttempt } from 'data/persistence/state/intrinsic';
+import { evalActivityAttempt, writePageAttemptState } from 'data/persistence/state/intrinsic';
 import { check } from '../../../../../../adaptivity/rules-engine';
 import {
   ApplyStateOperation,
   bulkApplyState,
   defaultGlobalEnv,
-  getEnvState
+  getEnvState,
 } from '../../../../../../adaptivity/scripting';
 import { createActivityAttempt } from '../../attempt/actions/createActivityAttempt';
 import { selectAll, selectExtrinsicState, setExtrinsicState } from '../../attempt/slice';
 import {
   selectCurrentActivityTree,
-  selectCurrentActivityTreeAttemptState
+  selectCurrentActivityTreeAttemptState,
 } from '../../groups/selectors/deck';
-import { selectPreviewMode, selectSectionSlug } from '../../page/slice';
+import { selectPreviewMode, selectResourceAttemptGuid, selectSectionSlug } from '../../page/slice';
 import { AdaptivitySlice, setLastCheckResults, setLastCheckTriggered } from '../slice';
 
 export const triggerCheck = createAsyncThunk(
@@ -24,9 +24,7 @@ export const triggerCheck = createAsyncThunk(
     const rootState = getState() as RootState;
     const isPreviewMode = selectPreviewMode(rootState);
     const sectionSlug = selectSectionSlug(rootState);
-    const currentActivityTreeAttempts = selectCurrentActivityTreeAttemptState(rootState) || [];
-    const currentAttempt = currentActivityTreeAttempts[currentActivityTreeAttempts?.length - 1];
-    const currentActivityAttemptGuid = currentAttempt?.attemptGuid || '';
+    const resourceAttemptGuid = selectResourceAttemptGuid(rootState);
 
     /* console.log('TRIGGER CHECK', {currentAttempt}); */
 
@@ -72,7 +70,21 @@ export const triggerCheck = createAsyncThunk(
         value: currentAttemptNumber + 1,
       },
     ];
-
+    let globalSnapshot = getEnvState(defaultGlobalEnv);
+    const isActivityAlreadyVisited = globalSnapshot[`${currentActivity.id}|visitTimestamp`];
+    // don't update the time if student is revisiting that page
+    if (!isActivityAlreadyVisited) {
+      //looks like SS captures the date when we leave the page so we will capture the time here for tracking history
+      // update the scripting
+      const targetVisitTimeStampOp: ApplyStateOperation = {
+        target: `${currentActivity.id}|visitTimestamp`,
+        operator: '=',
+        value: Date.now(),
+      };
+      updateScripting.push(targetVisitTimeStampOp);
+      // update the store
+      modifiedExtrinsicState[`${currentActivity.id}|visitTimestamp`] = Date.now();
+    }
     bulkApplyState(updateScripting, defaultGlobalEnv);
 
     //update the store with the latest changes
@@ -114,13 +126,22 @@ export const triggerCheck = createAsyncThunk(
     });
 
     const snapshot = getEnvState(defaultGlobalEnv);
-    const globalSnapshot = Object.keys(snapshot).reduce((collect: any, key: string) => {
+    globalSnapshot = Object.keys(snapshot).reduce((collect: any, key: string) => {
       if (key.indexOf('app.') === 0 || key.indexOf('variables.') === 0) {
         collect[key] = snapshot[key];
       }
       return collect;
     }, {});
 
+    if (!isPreviewMode) {
+      // update the server with the latest changes
+      /* console.log('trigger check last min extrinsic state', {
+        sectionSlug,
+        resourceAttemptGuid,
+        modifiedExtrinsicState,
+      }); */
+      await writePageAttemptState(sectionSlug, resourceAttemptGuid, modifiedExtrinsicState);
+    }
     await dispatch(setExtrinsicState({ state: modifiedExtrinsicState }));
     const stateSnapshot = {
       ...allResponseState,
@@ -148,22 +169,54 @@ export const triggerCheck = createAsyncThunk(
         stateSnapshot,
       }); */
     } else {
-      /* console.log('CHECKING', { sectionSlug, currentActivityTreeAttempts }); */
+      const currentActivityTreeAttempts = selectCurrentActivityTreeAttemptState(rootState) || [];
+      const currentAttempt = currentActivityTreeAttempts[currentActivityTreeAttempts?.length - 1];
+      const currentActivityAttemptGuid = currentAttempt?.attemptGuid || '';
+      /* console.log('CHECKING', {
+        sectionSlug,
+        currentActivityTreeAttempts,
+        currentAttempt,
+        currentActivityTree,
+      }); */
 
       if (!currentActivityAttemptGuid) {
         console.error('not current attempt, cannot eval', { currentActivityTreeAttempts });
         return;
       }
 
+      // BS: not sure why the current attempt responses are not up to date currently,
+      // for now just merge the tree state into the current attempt
+      // the LAYER is always up to date, but the current attempt is not for some reason
+      // this only occurs with layers, more investigation needed
+
       // we have to send all the current activity attempt state to the server
       // because the server doesn't know the current sequence id and will strip out
       // all sequence ids from the path for these only
       const partResponses: PartResponse[] =
-        currentAttempt?.parts.map(({attemptGuid, response}) => {
+        currentAttempt?.parts.map(({ partId, attemptGuid, response }) => {
+          // doing in reverse so that the layer's choice is the last one
+          const combinedResponse = currentActivityTreeAttempts.reverse().reduce(
+            (collect: any, attempt: any) => {
+              const part = attempt.parts.find((p: any) => p.partId === partId);
+              if (part) {
+                /* if (partId === 'orrery') {
+                  console.log('collecting parts:', { part, attempt, response, pr: part.response });
+                } */
+                if (part.response) {
+                  collect = {...collect, ...part.response};
+                }
+              }
+              return collect;
+            },
+            {},
+          );
+          const finalResponse = Object.keys(combinedResponse).length > 0 ? combinedResponse : null;
           // response should be wrapped in input, but only once
-          const input_response = response?.input ? response : { input: response };
-          return { attemptGuid, response: input_response };
+          /* const input_response = response?.input ? response : { input: response }; */
+          return { attemptGuid, response: { input: finalResponse } };
         }) || [];
+
+      /* console.log('PART RESPONSES', { partResponses, allResponseState }); */
 
       const evalResult = await evalActivityAttempt(
         sectionSlug,
@@ -175,15 +228,20 @@ export const triggerCheck = createAsyncThunk(
       isCorrect = checkResult.every((action: any) => action.params.correct);
     }
 
+    const currentActivityTreeAttempts = selectCurrentActivityTreeAttemptState(rootState) || [];
+    const currentAttempt = currentActivityTreeAttempts[currentActivityTreeAttempts?.length - 1];
+    const currentActivityAttemptGuid = currentAttempt?.attemptGuid || '';
     let attempt: any = currentAttempt;
     if (!isCorrect) {
       /* console.log('Incorrect, time for new attempt'); */
-      const {payload: newAttempt} = await dispatch(
+      const { payload: newAttempt } = await dispatch(
         createActivityAttempt({ sectionSlug, attemptGuid: currentActivityAttemptGuid }),
       );
       attempt = newAttempt;
     }
 
-    await dispatch(setLastCheckResults({ timestamp: currentTriggerStamp, results: checkResult, attempt }));
+    await dispatch(
+      setLastCheckResults({ timestamp: currentTriggerStamp, results: checkResult, attempt }),
+    );
   },
 );